Update override-http-client-adapter.md
Hi Frank! I think I've found an error on the docs so I try to propose a change. In fact, with the latest version of flutter data, it seems that with your previous setup the client get used more than one time (because dispose was never called) causing some trouble to all the callers except the first one.